Removed compiler warnings about signed/unsigned comparision.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/back/trunk@16943 72102866-910b-0410-8b05-ffd578937521

+++ Source/xlib/XGGState.m
@@ -914,7 +914,7 @@ static	Region	emptyRegion;
     {
       ll.y = y;
     }
-  if (ur.y > y + height)
+  if (ur.y > y + (int)height)
     {
       ur.y = y + height;
     }
